+++ upcloud/resource_upcloud_managed_database_test.go
@@ -1,6 +1,7 @@
 package upcloud
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/utils"
@@ -8,6 +9,10 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
+
+func withPrefix(text string) string {
+	return fmt.Sprintf("tf-acc-test-db-%s", text)
+}
 
 func TestAccUpcloudManagedDatabase(t *testing.T) {
 	testDataS1 := utils.ReadTestDataFile(t, "testdata/upcloud_managed_database/managed_database_s1.tf")
@@ -31,9 +36,9 @@ func TestAccUpcloudManagedDatabase(t *testing.T) {
 			{
 				Config: testDataS1,
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr(pg1Name, "name", "tf-pg-test-1"),
+					resource.TestCheckResourceAttr(pg1Name, "name", withPrefix("pg-1")),
 					resource.TestCheckResourceAttr(pg1Name, "plan", "1x1xCPU-2GB-25GB"),
-					resource.TestCheckResourceAttr(pg1Name, "title", "tf-test-pg-1"),
+					resource.TestCheckResourceAttr(pg1Name, "title", withPrefix("pg-1")),
 					resource.TestCheckResourceAttr(pg1Name, "zone", "pl-waw1"),
 					resource.TestCheckResourceAttr(pg1Name, "powered", "true"),
 					resource.TestCheckResourceAttr(pg1Name, "maintenance_window_time", "10:00:00"),
@@ -42,21 +47,24 @@ func TestAccUpcloudManagedDatabase(t *testing.T) {
 					resource.TestCheckResourceAttr(pg1Name, "properties.0.version", "13"),
 					resource.TestCheckResourceAttr(pg1Name, "type", string(upcloud.ManagedDatabaseServiceTypePostgreSQL)),
 					resource.TestCheckResourceAttrSet(pg1Name, "service_uri"),
+					resource.TestCheckResourceAttr(pg1Name, "network.#", "0"),
 
-					resource.TestCheckResourceAttr(pg2Name, "name", "tf-pg-test-2"),
+					resource.TestCheckResourceAttr(pg2Name, "name", withPrefix("pg-2")),
 					resource.TestCheckResourceAttr(pg2Name, "plan", "1x1xCPU-2GB-25GB"),
-					resource.TestCheckResourceAttr(pg2Name, "title", "tf-test-pg-2"),
+					resource.TestCheckResourceAttr(pg2Name, "title", withPrefix("pg-2")),
 					resource.TestCheckResourceAttr(pg2Name, "zone", "pl-waw1"),
 					resource.TestCheckResourceAttr(pg2Name, "powered", "true"),
 					resource.TestCheckResourceAttr(pg2Name, "properties.0.version", "14"),
+					resource.TestCheckResourceAttr(pg2Name, "network.#", "1"),
 
-					resource.TestCheckResourceAttr(msql1Name, "name", "tf-mysql-test-2"),
+					resource.TestCheckResourceAttr(msql1Name, "name", withPrefix("mysql-1")),
 					resource.TestCheckResourceAttr(msql1Name, "plan", "1x1xCPU-2GB-25GB"),
-					resource.TestCheckResourceAttr(msql1Name, "title", "tf-test-msql-1"),
+					resource.TestCheckResourceAttr(msql1Name, "title", withPrefix("mysql-1")),
 					resource.TestCheckResourceAttr(msql1Name, "zone", "pl-waw1"),
 					resource.TestCheckResourceAttr(msql1Name, "powered", "true"),
+					resource.TestCheckResourceAttr(msql1Name, "network.#", "0"),
 
-					resource.TestCheckResourceAttr(lgDBName, "name", "tf-test-logical-db-1"),
+					resource.TestCheckResourceAttr(lgDBName, "name", withPrefix("logical-db-1")),
 					resource.TestCheckResourceAttrSet(lgDBName, "service"),
 
 					resource.TestCheckResourceAttr(userName1, "username", "somename"),
@@ -74,17 +82,18 @@ func TestAccUpcloudManagedDatabase(t *testing.T) {
 					resource.TestCheckResourceAttr(userName4, "opensearch_access_control.0.rules.0.index", ".opensearch-observability"),
 					resource.TestCheckResourceAttr(userName4, "opensearch_access_control.0.rules.0.permission", "admin"),
 
-					resource.TestCheckResourceAttr(redisName, "name", "tf-redis-test-1"),
+					resource.TestCheckResourceAttr(redisName, "name", withPrefix("redis-1")),
 					resource.TestCheckResourceAttr(redisName, "plan", "1x1xCPU-2GB"),
-					resource.TestCheckResourceAttr(redisName, "title", "tf-test-redis-1"),
+					resource.TestCheckResourceAttr(redisName, "title", withPrefix("redis-1")),
 					resource.TestCheckResourceAttr(redisName, "zone", "pl-waw1"),
 					resource.TestCheckResourceAttr(redisName, "powered", "true"),
+					resource.TestCheckResourceAttr(redisName, "network.#", "1"),
 				),
 			},
 			{
 				Config: testDataS2,
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr(pg1Name, "title", "tf-test-updated-pg-1"),
+					resource.TestCheckResourceAttr(pg1Name, "title", withPrefix("pg-1-updated")),
 					resource.TestCheckResourceAttr(pg1Name, "maintenance_window_time", "11:00:00"),
 					resource.TestCheckResourceAttr(pg1Name, "maintenance_window_dow", "thursday"),
 					resource.TestCheckResourceAttr(pg1Name, "properties.0.public_access", "false"),
@@ -92,13 +101,15 @@ func TestAccUpcloudManagedDatabase(t *testing.T) {
 					resource.TestCheckResourceAttr(pg1Name, "properties.0.version", "14"),
 					resource.TestCheckResourceAttr(pg1Name, "powered", "false"),
 
-					resource.TestCheckResourceAttr(pg2Name, "title", "tf-test-updated-pg-2"),
+					resource.TestCheckResourceAttr(pg2Name, "title", withPrefix("pg-2-updated")),
 					resource.TestCheckResourceAttr(pg2Name, "powered", "true"),
 					resource.TestCheckResourceAttr(pg2Name, "properties.0.version", "14"),
+					resource.TestCheckResourceAttr(pg2Name, "network.#", "1"),
 
-					resource.TestCheckResourceAttr(msql1Name, "title", "tf-test-updated-msql-1"),
+					resource.TestCheckResourceAttr(msql1Name, "title", withPrefix("mysql-1-updated")),
+					resource.TestCheckResourceAttr(msql1Name, "network.#", "1"),
 
-					resource.TestCheckResourceAttr(lgDBName, "name", "tf-test-updated-logical-db-1"),
+					resource.TestCheckResourceAttr(lgDBName, "name", withPrefix("logical-db-1-updated")),
 
 					resource.TestCheckResourceAttr(userName1, "password", "Superpass890"),
 					resource.TestCheckResourceAttr(userName1, "authentication", "caching_sha2_password"),
@@ -110,9 +121,9 @@ func TestAccUpcloudManagedDatabase(t *testing.T) {
 					resource.TestCheckResourceAttr(userName3, "redis_access_control.0.commands.#", "0"),
 					resource.TestCheckResourceAttr(userName3, "redis_access_control.0.keys.0", "key*"),
 
-					resource.TestCheckResourceAttr(redisName, "name", "tf-redis-test-1"),
+					resource.TestCheckResourceAttr(redisName, "name", withPrefix("redis-1")),
 					resource.TestCheckResourceAttr(redisName, "plan", "1x1xCPU-2GB"),
-					resource.TestCheckResourceAttr(redisName, "title", "tf-test-redis-title-1"),
+					resource.TestCheckResourceAttr(redisName, "title", withPrefix("redis-1-updated")),
 					resource.TestCheckResourceAttr(redisName, "zone", "pl-waw1"),
 					resource.TestCheckResourceAttr(redisName, "powered", "true"),
 				),